requirements.txt
requests
beautifulsoup4
linetool
python-dotenv